PDA ping handler
Updated ping handler such that the user can disable it for a time, after
which it is automatically reactivated, while also allowing the user to
reactivate the handler before this time period has passed.

This allows the user to 'end a journey' without actually terminating the
program (as was the case in earlier versions), which would require the
user to start the application again before boarding a vehicle.